Remove useConfirmationRemoveModal hook
