+++ src/components/boiler-house-details/boiler-house-details.tsx
@@ -1,10 +1,9 @@
-import React from "react";
+import React, { useState } from "react";
 import { Link as RouterLink } from "react-router-dom";
 
 import { locale } from "../../services";
 import { ConfirmationModal } from "./confirmation-modal";
 import { useBoilerHouseDetails } from "./hooks/useBoilerHouseDetails";
-import { useConfirmRemoveModal } from "./hooks/useConfirmRemoveModal";
 
 import { Asset } from "@mtfh/common/lib/api/asset/v1";
 import { Button, Center, Heading, Link, Spinner } from "@mtfh/common/lib/components";
@@ -17,20 +16,14 @@ export const BoilerHouseDetails = ({ asset }: Props) => {
   const { isLoading, boilerHouseAsset, assetHasBoilerHouse } =
     useBoilerHouseDetails(asset);
 
-  const {
-    isLoading: isLoadingRemoveAssetRequest,
-    setShowModal,
-    showModal,
-    handleRemoveBoilerHouse,
-  } = useConfirmRemoveModal(asset);
+  const [showModal, setShowModal] = useState(false);
 
   return (
     <aside className="mtfh-cautionary-alerts">
       <ConfirmationModal
+        asset={asset}
         showModal={showModal}
-        hideModal={() => setShowModal(false)}
-        onSubmit={handleRemoveBoilerHouse}
-        isLoading={isLoadingRemoveAssetRequest}
+        setShowModal={setShowModal}
       />
 
       <Heading variant="h2" className="lbh-heading lbh-heading-h3">
